% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_BJ_bounds.R
\name{BJ_func}
\alias{BJ_func}
\title{Internal function - the BJ objective function minus the observed value, put it into uniroot to find the
test statistic magnitude bound at order statistic k of d that will give the observed BJ value.}
\usage{
BJ_func(x, k, d, b)
}
\description{
Internal function - the BJ objective function minus the observed value, put it into uniroot to find the
test statistic magnitude bound at order statistic k of d that will give the observed BJ value.
}
\keyword{internal}
