% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_b1.R
\name{calcb1}
\alias{calcb1}
\title{calc_b1.R}
\usage{
calcb1(lower = TRUE, upper = FALSE, muVec, sigMat, bounds)
}
\arguments{
\item{lower}{Boolean, whether to calculate lower bound.}

\item{upper}{Boolean, whether to calculate upper bound.}

\item{muVec}{Mean vector of MVN (under the alternative).}

\item{sigMat}{Covariance matrix of MVN (under the alternative).}

\item{bounds}{A d*1 vector of bounds on the magnitudes of the test statistics, where
the first element is the bound for |Z|_(1) and the last element is the bound for |Z|_(d).
If provided, will calculate power from crossing the bounds.}
}
\value{
A list with the elements:
\item{allProbsLower}{J*1 vector of all component probabilities.}
\item{lowerProb}{Lower bound.}
\item{allProbsUpper}{J*1 vector of all component probabilities.}
\item{upperProb}{Upper bound.}
}
\description{
Calculate lower bound 1 - R1 or upper bound 1 - S1.
}
\examples{
myVariance <- matrix(data=0.3, nrow=5, ncol=5)
diag(myVariance) <- 1
myBounds <- set_GBJ_bounds(alpha = 0.01, d=5, sig_vec = myVariance[lower.tri(myVariance)])
calcb1(muVec = c(1, 0, 0, 0, 0), sigMat = myVariance, bounds=myBounds)

}
