% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_mvn.R
\name{checkBoundsCross}
\alias{checkBoundsCross}
\title{Internal function to check if bounds are violated.
Bounds should be smallest to largest.}
\usage{
checkBoundsCross(x, bounds)
}
\description{
Internal function to check if bounds are violated.
Bounds should be smallest to largest.
}
\keyword{internal}
