% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_b1.R
\name{checkBoundsLower1}
\alias{checkBoundsLower1}
\title{Internal function to check if test statistics fall within the region
used for b1 lower bound, 1 minus the result of this is the lower bound on the power..}
\usage{
checkBoundsLower1(x, bounds)
}
\description{
Internal function to check if test statistics fall within the region
used for b1 lower bound, 1 minus the result of this is the lower bound on the power..
}
\keyword{internal}
