% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_b2.R
\name{checkBoundsLower2}
\alias{checkBoundsLower2}
\title{Internal function to check if test statistics fall within the region
used for b2 lower bound, 1 minus the result of this is the lower bound on the power.}
\usage{
checkBoundsLower2(x, J, bounds)
}
\description{
Internal function to check if test statistics fall within the region
used for b2 lower bound, 1 minus the result of this is the lower bound on the power.
}
\keyword{internal}
