% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_b1.R
\name{checkBoundsUpper1}
\alias{checkBoundsUpper1}
\title{Internal function to check if test statistics fall within the region
used for b1 uupper bound, 1 minus the result of this is the upper bound on the power.}
\usage{
checkBoundsUpper1(x, J, bounds)
}
\description{
Internal function to check if test statistics fall within the region
used for b1 uupper bound, 1 minus the result of this is the upper bound on the power.
}
\keyword{internal}
