% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_b2.R
\name{createMjk}
\alias{createMjk}
\title{Create the matrix that linearly transforms the vector of test statistics
into a quantity amenable for pmvnorm.}
\usage{
createMjk(j, k, size)
}
\arguments{
\item{j}{The element of the vector that is the largest.}

\item{k}{The element of the vector that is the second largest.}

\item{size}{The length of the set.}
}
\value{
The transformation matrix of dimension (2J-1)*(2J-1)
}
\description{
Create the matrix that linearly transforms the vector of test statistics
into a quantity amenable for pmvnorm.
}
\examples{
createMjk(j=3, k=4, size=5)
}
