% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_GBJ_bounds.R
\name{GBJ_bound1_root}
\alias{GBJ_bound1_root}
\title{Internal function - put this function into uniroot to find the magnitude of the largest test statistic that will
give a p-value of alpha.}
\usage{
GBJ_bound1_root(x, alpha, d, sig_vec)
}
\description{
Internal function - put this function into uniroot to find the magnitude of the largest test statistic that will
give a p-value of alpha.
}
\keyword{internal}
