% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_b2.R
\name{performIntegralUpper2}
\alias{performIntegralUpper2}
\title{Apply this function over all m, j not equal (order matters) to calculate each portion of the integral
we need for the lower bound for calc_b2.}
\usage{
performIntegralUpper2(
  x,
  muVec,
  sigMat,
  lBounds1,
  uBounds1,
  lBounds2,
  uBounds2,
  lBounds3,
  uBounds3,
  lBounds4,
  uBounds4
)
}
\arguments{
\item{x}{Apply over this 2*1 vector, the elements that will be the largest and second largest in magnitude.}

\item{muVec}{Mean vector of test statistics under the alternative (assuming it's MVN).}

\item{sigMat}{Covariance matrix of test statistics under the alternative (assuming it's MVN).}

\item{lBounds1}{A 3J-2 vector of lower bounds for the first integral (see paper).}

\item{uBounds1}{A 3J-2 vector of upper bounds for the second integral (see paper).}

\item{lBounds2}{A 3J-2 vector of lower bounds for the first integral (see paper).}

\item{uBounds2}{A J3J-2 vector of upper bounds for the second integral (see paper).}

\item{lBounds3}{A 3J-2 vector of lower bounds for the third integral (see paper).}

\item{uBounds3}{A 3J-2 vector of upper bounds for the third integral (see paper).}

\item{lBounds4}{A 3J-2 vector of lower bounds for the fourth integral (see paper).}

\item{uBounds4}{A 3J-2 vector of upper bounds for the fourth integral (see paper).}
}
\value{
The value of the integration.
}
\description{
Apply this function over all m, j not equal (order matters) to calculate each portion of the integral
we need for the lower bound for calc_b2.
}
