% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_from_bounds.R
\name{power_from_bounds}
\alias{power_from_bounds}
\title{power_from_bounds.R}
\usage{
power_from_bounds(bounds, sig_mat, muVec)
}
\arguments{
\item{bounds}{A d*1 vector of bounds on the magnitudes of the test statistics, where
the first element is the bound for |Z|_(1) and the last element is the bound for |Z|_(d).}

\item{sig_mat}{The covariance matrix of the test statistics (assume multivariate normal).}

\item{muVec}{The mean vector of the test statistics (assume multivariate normal).}
}
\value{
Power of the test.
}
\description{
For detection boundary type tests, find the power given the rejection region bounds and
type of alternative
}
\examples{
myVariance <- matrix(data=0.3, nrow=5, ncol=5)
diag(myVariance) <- 1
myBounds <- set_GBJ_bounds(alpha = 0.01, d=5, sig_vec = myVariance[lower.tri(myVariance)])
power_from_bounds(bounds = myBounds, sig_mat = myVariance, muVec = c(1, 0, 0, 0, 0))

}
