% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_lower_bound.R
\name{sim_lower_bound}
\alias{sim_lower_bound}
\title{sim_lower_bound.R}
\usage{
sim_lower_bound(n, dim, muVec, sigMat, bounds)
}
\arguments{
\item{n}{Number of simulations.}

\item{dim}{1 for R1 or 2 for R2}

\item{muVec}{Mean vector of MVN (under the alternative).}

\item{sigMat}{Covariance matrix of MVN (under the alternative).}

\item{bounds}{A d*1 vector of bounds on the magnitudes of the test statistics, where
the first element is the bound for |Z|_(1) and the last element is the bound for |Z|_(d).
If provided, will calculate power from crossing the bounds.}
}
\value{
Lower bound on the probability.
}
\description{
Simulate the lower bounds R1 and R2.
}
\examples{
myVariance <- matrix(data=0.3, nrow=5, ncol=5)
diag(myVariance) <- 1
myBounds <- set_GBJ_bounds(alpha = 0.01, d=5, sig_vec = myVariance[lower.tri(myVariance)])
sim_lower_bound(n=5000, muVec = c(1, 0, 0, 0, 0), sigMat = myVariance, bounds=myBounds)

}
