% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_mvn.R
\name{sim_power_mvn}
\alias{sim_power_mvn}
\title{sim_power_mvn.R}
\usage{
sim_power_mvn(
  n,
  muVec,
  sigMat,
  nullSigMat = NULL,
  bounds = NULL,
  test = NULL,
  alpha
)
}
\arguments{
\item{n}{Number of simulations.}

\item{muVec}{Mean vector of MVN (under the alternative).}

\item{sigMat}{Covariance matrix of MVN (under the alternative).}

\item{nullSigMat}{Assumed correlation matrix of MVN under the null.}

\item{bounds}{A d*1 vector of bounds on the magnitudes of the test statistics, where
the first element is the bound for |Z|_(1) and the last element is the bound for |Z|_(d).
If provided, will calculate power from crossing the bounds.}

\item{test}{Either "GHC", "HC", "GBJ", or "BJ" or NULL. If provided, will calculate
the p-value using the specified test and calculate power this way.}

\item{alpha}{Level of the test}
}
\value{
A list with the elements:
\item{boundsPower}{Power from using bounds approach.}
\item{testPower}{Power from using specific test p-value approach.}
}
\description{
Simulate power of detection boundary tests starting from multivariate normal test statistics.
}
\examples{
myVariance <- matrix(data=0.3, nrow=5, ncol=5)
diag(myVariance) <- 1
myBounds <- set_GBJ_bounds(alpha = 0.01, d=5, sig_vec = myVariance[lower.tri(myVariance)])
sim_power_mvn(n=1000, muVec = c(1, 0, 0, 0, 0), sigMat = myVariance, alpha=0.01)

}
