% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_R1c.R
\name{sim_R1c}
\alias{sim_R1c}
\title{sim_R1c.R}
\usage{
sim_R1c(n, muVec, sigMat, bounds)
}
\arguments{
\item{n}{Number of simulations.}

\item{muVec}{Mean vector of MVN (under the alternative).}

\item{sigMat}{Covariance matrix of MVN (under the alternative).}

\item{bounds}{A d*1 vector of bounds on the magnitudes of the test statistics, where
the first element is the bound for |Z|_(1) and the last element is the bound for |Z|_(d).
If provided, will calculate power from crossing the bounds.}
}
\value{
Lower bound on the probability.
}
\description{
Simulate the upper bound R1c.
}
\examples{
myVariance <- matrix(data=0.3, nrow=5, ncol=5)
diag(myVariance) <- 1
myBounds <- set_GBJ_bounds(alpha = 0.01, d=5, sig_vec = myVariance[lower.tri(myVariance)])
sim_R1c(n=5000, muVec = c(1, 0, 0, 0, 0), sigMat = myVariance, bounds=myBounds)

}
